+++ lib/devbox_launcher.rb
@@ -3,6 +3,10 @@ require "ssh-config"
 require "open3"
 require "thor"
 require "net/ssh"
+require "os"
+require "ruby-watchman"
+require 'socket'
+require 'pathname'
 require "yaml"
 require "devbox_launcher/version"
 
@@ -11,3 +15,4 @@ module DevboxLauncher
 end
 
 require "devbox_launcher/cli"
+require "devbox_launcher/watchman"