+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0]
+### Changed
+- Label sessions with devbox
+- On linux, mutagen relies on watchman to detect changes
+
+### Fixed
+- In Linux, mutagen no longer burns CPU every 10 seconds
+
+## [0.2.0]
 ### Added
 - Initial release